Made RefCounted and ThreadSafeRefCounted not have virtual destructors, as the virtual destructors were making some classes polymorphic (have a vtable) that didn't need to be. Suppressing some padding warnings.
git-svn-id: file:///home/nick/svnrepo/repo/indigo/trunk@10768 20aa11e4-6e6f-594c-a032-6291ace9ee65

+++ utils/CircularBuffer.cpp
@@ -441,7 +441,7 @@ void circularBufferTest()
 						buf.clear();
 					}
 
-					testAssert(test_struct->getRefCount() == 1 + buf.size());
+					testAssert(test_struct->getRefCount() == (int64)(1 + buf.size()));
 
 					// Test iterating over resulting buffer
 					size_t c = 0;
